Reconfigured Stripe backend
