+++ client/package.json
@@ -34,6 +34,7 @@
     "react-spinners-kit": "^1.9.0",
     "react-stockcharts": "^0.7.8",
     "react-stripe-checkout": "^2.6.3",
+    "react-stripe-elements": "^2.0.3",
     "react-with-firebase-auth": "^1.0.2",
     "recompose": "^0.30.0",
     "redux": "^4.0.1",
