--- client/src/constants/payment_server.js
@@ -1,5 +1,0 @@
-const PAYMENT_SERVER_URL = process.env.NODE_ENV === 'production'
-  ? 'https://getpickem.co'
-  : 'http://localhost:5000';
-
-export default PAYMENT_SERVER_URL;
