+++ client/src/constants/stripe.js
@@ -1,5 +1,5 @@
 const STRIPE_PUBLISHABLE = process.env.NODE_ENV === 'production'
   ? 'pk_test_7mDiIF0ntVgLJjKz6B7evL0i00J88bDjzD'
-  : 'pk_test_7mDiIF0ntVgLJjKz6B7evL0i00J88bDjzD'
+  : 'pk_test_7mDiIF0ntVgLJjKz6B7evL0i00J88bDjzD';
 
- export default STRIPE_PUBLISHABLE;
+export default STRIPE_PUBLISHABLE;
