--- server/routers/index.js
@@ -1,7 +1,0 @@
-const paymentApi = require('./payment');
-
-const configureRoutes = app => {
-  paymentApi(app);
-};
-
-module.exports = configureRoutes;
